Issue#146: Pass the name of the lib to load at runtime
This is done not as environment vars but as a literal that can be stringified
at compilation time. This ensures that the version of the lib that we load
to work around some simulators not doing global symbol resolution matches
that which was used as build time. If we used name=get_env("") this would
be more prone to problems.