Issue #17:  Various changes for vhpi iteration
1. Some VHPI implementations do not allow creation of
and iterator of arbitary type on a handle. The architecture
inheritance diagrams seem to specify what the allowed
one-2-many relationships are for a specific handle type. We
now have mapping of lists that can be used for each handle type

2. Rather than depond on the length of a BinaryValue to determine
the number of children a handle has call a new function
get_num_elems that can make an informed decision.