Issue #17: Allow the next_handle to express more than PASS/FAIL
The previous GpiIterator::next_handle culd only express PASS/FAIL
but we really needed to know if an object had been found
but creation failed. The main reason for this is that the object
was visible but of a different language. We now allow the function
to return if this is the case the GPI layer can take appropriate
action.