Issue #17: Handle boolean values.
This should have been easy but showed an interesting issue with
vhpi. Both std_logic and boolean end up appearing as vhpiEnumTypeDeclK
but the method used to assign data is different. std_logic needing to
use vhpi[X,U,Z,0,1] and boolean needing vhpi[True,False]. If this
is not done then there is a range violation on the boolean access.

Delving into the BaseType and qurying the number of literals that
a signal has gives a likley probablity that a signal is boolean or
not.

Adding a subclass of vhpiSignalObjHdl then allows a specific
set_signal_value to be used for the two cases.

This is probably also a pre-cursor to using the base type
rather than the format to selet the type of object to create