Correcting usage of SIM_BUILD for ius scratch dir
Here's an excerpt from Cadence docs regarding the usage of nclibdirname
and nclibdirpath.

Changing the Name of the INCA_libs Directory

By default, irun generates a scratch directory during simulation called
INCA_libs and saves it to the current working directory. Use the
-nclibdirname option to change the name.

For example:

% irun -nclibdirname IRUN_libs dut.v tb.v

This command changes the name of the scratch directory to IRUN_libs,
which is generated during simulation and saved to the current working
directory by irun.

Changing the Path to a Custom Scratch Directory

You can include path information when specifying a custom name for the
INCA_libs scratch directory. Path statements passed to the -nclibdirname
command-line option can be relative or absolute, as shown:

% irun -nclibdirname ../IRUN_libs dut.v tb.v

In this example irun is used to change the name of the scratch directory
to IRUN_libs, but instead of saving it to the current working directory,
the tool will save the directory using a relative path that is located
above the current working directory.

% irun -nclibdirname ./dut/IRUN_libs dut.v tb.v

This example specifies a relative path in the current working directory
with multiple hierarchical levels. The directory ./dut must exist in
order for irun to create the new IRUN_libs location. If ./dut does not
exist, irun will halt and generate an NWRKDRA error.

Changing the Path to the INCA_libs Directory

Alternatively, you can use the -nclibdirpath option on the irun command
line when specifying a relative path for storing new libraries. This
option can be used by itself or together with together with
-nclibdirname.

When using -nclibdirname together with -nclibdirpath, the directory name
specified with -nclibdirname must not contain any path information.
For example:

% irun -nclibdirpath ./libraries dut.v tb.v

This irun command will generate a new INCA_libs directory in the
relative location ./libraries, which is located in the current working
directory. Note that ./libraries must exist in order to avoid an error.

% irun -nclibdirname IRUN_libs -nclibdirbath ../libraries dut.v tb.v

This irun command will generate a new IRUN_libs directory in the
relative location ../libraries, which is one level above the current
working directory. Note that ../libraries must exist in order to avoid
an error. The tool will automatically generate IRUN_libs, if
../libraries is available.