Add labels to all str.format calls to fix Python 2.6 build
In Python 2.6 and below, a label (i.e. 0, 1, etc.) is necessary
in calls to str.format(). That is, {} is not valid in 2.6.x;
instead, it needs to be {0}. There were a handful of places in
the codebase using format() without such labels. This commit
adds the relevant labels.

The cocotb documentation claims Python 2.6+ is still supported.
Without this change, when ran against Python 2.6, a "Called
callback returned null" message will be generated at the end of
simulation, when cocotb tries to print out the number of failed
tests and the consumed simulation time,

In particular, RHEL/SL/CentOS 6 is sadly still using Python 2.6,
so this change makes cocotb work with Python out-of-the-box on
those distros.