Fix for random crashes under Ubuntu 16.04 LTS.
Problem was caused by faulty pointer passing between C and Python. Under
Ubuntu 16.04 LTS, the C library alloc (or loader?) causes virtual
addresses to be used in the upper half of the 32 bit address space, when
a 32-bit simulator and 32-bit Python is used. The integer value in
Python can not be converted using an "Unsigned" conversion function in
that case. The resulting pointer is -1, which causes a SEGFAULT.