Updated lib ref rst to included implementations included in cocotb package.  Added info message to notify user when strict_type checking is on.
Added requested changes, moved the message about strict_type to the test_failure.

Touch up of the documentation once more. Added sphinx usage of the napoleon extension which should be available in a standard install of sphinx > 1.3

Added copyright update on build by importing the datetime module and grabbing the current year.  Added all the handles found in the handle module plus the SimHandle function to documentation.

Fixed a typo, added hooks to be autodoc'd, fixed a py class autodoc reference.