Minor fix to new test_singleton_isinstance in test_cocotb.py
This is cleaning up after a change made in #752. The NextTimeStep
trigger wasn't imported from cocotb.triggers, and also the "s" in
Timestep was not capitalized in the test itself.

I was testing the latest cocotb git master on ius (Cadence Incisive)
and happened to notice this. All the other tests seem to be passing,
which is good!