Move all installed files into the cocotb directory
setuptools depends on having all installed files below the cocotb
directory. Move the files there.

This commit is preparing to add Python packaging to cocotb, but should
not yet require any changes for people using cocotb by pointing the
COCOTB environment variable to the cocotb top-level source directory.

For backwards-compat two Makefiles,
$COCOTB/makefiles/Makefile.{sim,inc} which call the Makefile.{sim,inc}
at the new path.

Note that this commit does *not* add symlinks for the old directory
paths, as this approach wouldn't work on Windows.