Make USER_DIR independent of SIM_ROOT
USER_DIR is the path containing the user-called Makefile, i.e. where the
user code lives. Extract this path from MAKEFILE_LIST intead of relying
on SIM_ROOT.

This is preparing for a world where the cocotb directory structure and
with it SIM_ROOT will change. No user-visible changes are expected after
applying this commit.