Add Python packaging to cocotb
This patch adds a setup.py file creating a Python package for cocotb.
Most files are auto-declared as dependency and hence don't need to be
present in MANIFEST.in.

If cocotb is installed from a Python package, the VERSION environment
variable, containing the cocotb version, is now read from the installed
package and no longer from the version file in cocotb's top-level source
directory.

If cocotb is used from the source tree, the VERSION variable remains to
be read from the srcroot/version file as before.