Split SIM_ROOT to COCOTB_PY_DIR & COCOTB_SHARE_DIR
SIM_ROOT was defined to be the root of the source directory of cocotb,
and used in various places. Soon we will be moving code around in this
directory, moving the {include,lib,makefiles} directories into
cocotb/share. To prepare for this change this commit changes the code to
use two separate variables for the different directories.

COCOTB_PY_DIR refers to the directory containing the "cocotb"
subdirectory with the Python package. Add COCOTB_PY_DIR to PYTHONPATH to
use the cocotb Python code.

COCOTB_SHARE_DIR refers to the directory containing the
{include,lib,makefiles} directories. These will soon move to
cocotb/share in the next step, that's where the name comes from.

No change in behavior should be observed after applying this patch, it's
an internals-only change.