Added info to NullTrigger from Event.wait()
NullTriggers returned as a result of waiting on a fired Event now
include the Event name.

In the case of a looped coroutine waiting on an Event that has been
fired but not cleared, the exhibited behavior is usually hitting the
recursion limit in the scheduler and throwing an exception. Having the
information about the Event in the NullTrigger object is very helpful
when debugging.