Simplify the logging wrappers
It seems to me the only purpose of the `SimLog` object is to allow C to call into the python logging infrastructure via `_printRecord`.

However, that desire alone does not justify the need to create our own class - we can simply move the  `_printRecord` method into `SimLogBase` (renamed to `_logFromC`), and then none of rest of the wrapping is needed.

This also fixes some other problems:
* `SimBaseLog` did not call the right `__init__`, and re-initialized members already initialized by the base class
* `SimLog` did not support all of the features of the underlying logger (such as exception information)

Tested by looking at the travis log file before and after, and seeing nothing has changed