Use slots in triggers for speed and memory savings
Also remove some unnecessary attributes:
* `Trigger.signal` was meaningless, and exists already in `_EdgeBase.signal`
* `Join.pass_retval` is ignored by all of cocotb, and is a remnant of an older scheduler iteration

A side effect of this is it prevents users attaching extra properties to trigger instances - that's probably not an issue.
User-defined trigger subclasses (if any exist) will not be affected by this change, but could now opt into the same efficiency with their own `__slots__`.

Performance measurements due to this change will hopefully appear in the comments of this github PR.