Questa: start simulation from PWD, not PWD/sim_build.
The Icarus and Cadence (ius) simulator targets run the simulation
relative to PWD; i.e. the location of the Makefile. The Questa one,
though, seems to run everything out of the sim_build directory.

This commit changes the questa Makefile to run from ${PWD}, while
keeping the "work" library and the run script in sim_build, for better
compatibility with the other simulators.

In particular, I just tried to port a project from Cadence to Questa
and found that, because my Makefile didn't say ${PWD} but instead
just provided source paths relative to the current working directory,
the simulation would not start. Hence this change.

Hopefully this does not break anything: it would be great if people
with Questa projects could test!