Don't generate waveform for Verilator
The dump is not supported as Verilator supports a better mechanism
which is controlled by VERILATOR_TRACE in the Makefile.

Actually the best way is to put the initial with the dumps into a
seperate module and make both toplevel in all but Verilator, but that
needs proper testing on all simulators.