Verilator: Set precision at runtime
Verilator has no timing, so that a precision much below the actual
clock does not add value. On the other hand, it (currently) simulates
each cycle.

The optimal solution is to configure the precision at runtime to be
the gcd of all clocks or drive the clocks from the Verilator CPP
bench. Until this is supported, this patch adds support to configure
the precision properly in the Makefile. It should be set as near to
the clock as possible.