Use command(1) rather than which(1) in Makefiles
which(1) is not part of the POSIX standard, and is not present on
several systems and build environments out of the box (e.g. the
Nix sandbox). It's not actually part of GNU coreutils as many assume;
GNU has its own version (http://www.gnu.org/software/which/) but I
believe Debian might ship its own implementation.

`type -p` is a commonly-suggested alternative, but its output is
unfortunately not stable across systems. `command -v` is fully
portable across all POSIX systems.

We need to use $(shell :; command ...) because make tries to exec
programs directly rather than use the shell where possible, but the
logic for it is broken on old versions (omitting command(1) among other
shell builtins). This breaks things when command(1) is available only
as a shell builtin and not as an external command.

References:
* https://twitter.com/whitequark/status/1168670168173559808
* https://twitter.com/dev_console/status/1168672291271692288
* https://pubs.opengroup.org/onlinepubs/9699919799/utilities/command.html
* https://git.savannah.gnu.org/cgit/make.git/tree/job.c?h=4.2&id=ec6198012261e11b72bd5a95bd9e5c4d895c5f22#n2433