Provide cocotb environment variables for sim time
Simulators generally have an option to provide a timescale. Previously,
this could be set through the EXTRA_ARGS make variable. With this PR,
cocotb provides two new variables for exactly this purpose:
COCOTB_HDL_TIMEUNIT and COCOTB_HDL_TIMEPRECISION. The default is
"1ns/1ps". See the Verilog LRM for a full description on the
semenantics.