Do not ignore failures during test initialization. Fixes #1253.
There are two places in the RM that initialize coroutines, each had
slightly different code, so this was factored out into a method
init_test.

init_test can fail, and so can a test at any point in it's run
(previously handled by handle_result), so scoring a test as a pass or a
fail depending upon the result and expected behavior is factored out
into a method score_test.

Now results are handled during initialization, which fixes #1253.