Implemented PLUSARGS for FLI
The FLI does not offer any utility to easily grab the argc/argv from the
simulator. I use Questa's included TCL interpreter C API to access the
TCL 'argv' variable.

The 'argv' variable references the outer vsim call's arguments due to how
the current makefiles are setup, so the PLUSARGS variable was added back
to the outer call. PLUSARGS must also remain on the inner call for the
VPI implementation to work.