Use LogRecord.getMessage() to get format log message
In our customized loggers, we relied on manually creating a log message
string from the record.msg and record.args. Instead of doing so
manually, this commit uses LogRecord.getMessage(), which is the same
approach taken by Python's default Formatter class.

This change was motivated by the fact that `str()` was applied to log
messages when they went through the non-colored log formatter, but not
when going through the colored one, leading to an exception when writing
the log entry.

Fixes #1408